Move sleep display handling to custom_status_screen.c for peripheral support
The sleep event handler was in screen.c which is only compiled for
central/non-split builds. Since toucan_pet runs on the peripheral
(toucan_left), the sleep screen never displayed.

- Added activity state listener to custom_status_screen.c that swaps
  the LVGL image source between toucan128 and sleep_icon
- Moved sleep.c outside the split-role guard so the sleep icon data
  is available for all builds
- Made sleep_icon non-static so it can be referenced via LV_IMG_DECLARE
- Removed duplicate activity handler from screen.c (central builds)
