+++ boards/shields/toucan_pet/CMakeLists.txt
@@ -8,11 +8,11 @@ if(CONFIG_ZMK_DISPLAY AND CONFIG_NICE_VIEW_WIDGET_STATUS)
   zephyr_library_sources(widgets/output.c)
   zephyr_library_sources(widgets/util.c)
   zephyr_library_sources(widgets/toucan128.c)
+  zephyr_library_sources(widgets/sleep.c)
 
   if(NOT CONFIG_ZMK_SPLIT OR CONFIG_ZMK_SPLIT_ROLE_CENTRAL)
   zephyr_library_sources(widgets/layer.c)
   zephyr_library_sources(widgets/profile.c)
   zephyr_library_sources(widgets/screen.c)
-  zephyr_library_sources(widgets/sleep.c)
   endif()
 endif()
